+++ tests/losses/test_cosface_loss.py
@@ -14,7 +14,7 @@ class TestCosFaceLoss(unittest.TestCase):
         scale = 64
 
         for dtype in [torch.float16, torch.float32, torch.float64]:
-            loss_func = CosFaceLoss(margin=margin, scale=scale, num_classes=10, embedding_size=2, dtype=dtype).to(self.device)
+            loss_func = CosFaceLoss(margin=margin, scale=scale, num_classes=10, embedding_size=2).to(self.device)
 
             embedding_angles = torch.arange(0, 180)
             embeddings = torch.tensor([c_f.angle_to_coord(a) for a in embedding_angles], requires_grad=True, dtype=dtype).to(self.device) #2D embeddings
